mexico-decarb-scenario-explorer/requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
geopandas>=0.14.0
scipy>=1.11.0
pyyaml>=6.0
pytest>=7.4.0
openpyxl>=3.1.0
